+++ core/src/main/kotlin/cache/data/RegionData.kt
@@ -9,14 +9,13 @@ data class RegionData(
     val id: String,
     val guildId: OptionalSnowflake,
     val name: String,
-    val vip: Boolean,
     val optimal: Boolean,
     val deprecated: Boolean,
     val custom: Boolean,
 ) {
     companion object {
         fun from(guildId: OptionalSnowflake, region: DiscordVoiceRegion) = with(region) {
-            RegionData(id, guildId, name, vip, optimal, deprecated, custom)
+            RegionData(id, guildId, name, optimal, deprecated, custom)
         }
     }
 }