+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * Fixed the space in active threads get endpoint causing it to fail.
 * Fixed incorrect serialization of `NUMBER` option
+* Fixed Missing properties in `MemberData`, `InteractionData` and others. (#361)
+
+## Changes
+
+* `Region` no longer has a `vip` property 
 
 # 0.8.0-M3